refactor: speaker-state — absorb slider apply into the controller
Push the volume/bass/balance apply-merge (plus confirm) out of
speaker-state.js's FIELDS table and into the per-field slider controller.
The controller now owns the full user-driven slider arc end-to-end:
set, queue, flush, reconcile incoming WS values, confirm.

Move sliders.js up out of actions/ since it's no longer purely a command-
layer concern — it owns user-intent state that the WS reconciler reads
back through applyIncoming(state, value). speaker-state.js now has zero
imports from actions/, breaking the bidirectional coupling.

Drop the actions._confirmSlider export (no remaining external consumers)
and the per-field afterApply hook. hasPending stays — speaker-state tests
still assert it cross-module to pin the "confirm gates a follow-up set"
invariant.
